[enhancement] some code changes

+++ pytorch/test_model/efficient_module.py
@@ -540,7 +540,7 @@ def build_conv_layer(cfg, *args, **kwargs):
 
     layer = conv_layer(*args, **kwargs, **cfg_)
 
-    return layer
+    return layer.double()
 
 
 norm_cfg = {
